refactor: state, game 로직 분리
